requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.22.0
websockets>=11.0.3
httpx>=0.24.1
cryptography>=41.0.7
qrcode[pil]>=7.4.2
jinja2>=3.1.2
pydantic>=2.7.0
python-multipart>=0.0.6
brotli>=1.0.9
aiohttp>=3.8.0
pyjwt>=2.8.0
psutil>=5.9.0
pure-protobuf>=3.1.2
